feat(icon): convert icon components to ts
