--- packages/icon/src/index.js
@@ -1,4 +1,0 @@
-export { default } from './react/index.js'
-export * from './vars/index.js'
-export * from './react/icons/index.js'
-export { default as css } from './css/index.js'
