+++ packages/icon/src/index.tsx
@@ -1,0 +1,4 @@
+export { default } from './react/'
+export * from './vars/'
+export * from './react/icons/'
+export { default as css } from './css/'
