--- packages/icon/src/react/icons/index.js
@@ -1,163 +1,0 @@
-export { AccountIcon } from './AccountIcon.dist.js'
-export { AnalyticsIcon } from './AnalyticsIcon.dist.js'
-export { ArrowDownIcon } from './ArrowDownIcon.dist.js'
-export { ArrowDownRightIcon } from './ArrowDownRightIcon.dist.js'
-export { ArrowLeftIcon } from './ArrowLeftIcon.dist.js'
-export { ArrowRightIcon } from './ArrowRightIcon.dist.js'
-export { ArrowUpIcon } from './ArrowUpIcon.dist.js'
-export { ArrowUpRightIcon } from './ArrowUpRightIcon.dist.js'
-export { ArticleIcon } from './ArticleIcon.dist.js'
-export { AuthorKitIcon } from './AuthorKitIcon.dist.js'
-export { AuthorsNestIcon } from './AuthorsNestIcon.dist.js'
-export { BellIcon } from './BellIcon.dist.js'
-export { BellRungIcon } from './BellRungIcon.dist.js'
-export { BookmarkFillIcon } from './BookmarkFillIcon.dist.js'
-export { BookmarkIcon } from './BookmarkIcon.dist.js'
-export { BriefcaseIcon } from './BriefcaseIcon.dist.js'
-export { BrowseIcon } from './BrowseIcon.dist.js'
-export { BuildingIcon } from './BuildingIcon.dist.js'
-export { CalendarIcon } from './CalendarIcon.dist.js'
-export { CaretAltDownIcon } from './CaretAltDownIcon.dist.js'
-export { CaretAltLeftIcon } from './CaretAltLeftIcon.dist.js'
-export { CaretAltRightIcon } from './CaretAltRightIcon.dist.js'
-export { CaretAltUpIcon } from './CaretAltUpIcon.dist.js'
-export { CaretDownIcon } from './CaretDownIcon.dist.js'
-export { CaretLeftIcon } from './CaretLeftIcon.dist.js'
-export { CaretRightIcon } from './CaretRightIcon.dist.js'
-export { CaretUpIcon } from './CaretUpIcon.dist.js'
-export { ChannelAddIcon } from './ChannelAddIcon.dist.js'
-export { ChannelIcon } from './ChannelIcon.dist.js'
-export { ChatAltFillIcon } from './ChatAltFillIcon.dist.js'
-export { ChatAltIcon } from './ChatAltIcon.dist.js'
-export { ChatFillIcon } from './ChatFillIcon.dist.js'
-export { ChatIcon } from './ChatIcon.dist.js'
-export { CheckCircleFillIcon } from './CheckCircleFillIcon.dist.js'
-export { CheckCircleIcon } from './CheckCircleIcon.dist.js'
-export { CheckIcon } from './CheckIcon.dist.js'
-export { ClockIcon } from './ClockIcon.dist.js'
-export { ClockWarningIcon } from './ClockWarningIcon.dist.js'
-export { CloseCaptionedFillIcon } from './CloseCaptionedFillIcon.dist.js'
-export { CloseCaptionedIcon } from './CloseCaptionedIcon.dist.js'
-export { CloseIcon } from './CloseIcon.dist.js'
-export { CloudIcon } from './CloudIcon.dist.js'
-export { CloudOffIcon } from './CloudOffIcon.dist.js'
-export { CodeBrandedIcon } from './CodeBrandedIcon.dist.js'
-export { CodeIcon } from './CodeIcon.dist.js'
-export { ConnectionWarningIcon } from './ConnectionWarningIcon.dist.js'
-export { CopyIcon } from './CopyIcon.dist.js'
-export { CoursesIcon } from './CoursesIcon.dist.js'
-export { CurriculumIcon } from './CurriculumIcon.dist.js'
-export { DashboardIcon } from './DashboardIcon.dist.js'
-export { DesktopIcon } from './DesktopIcon.dist.js'
-export { DownloadIcon } from './DownloadIcon.dist.js'
-export { EnvelopeArrowIcon } from './EnvelopeArrowIcon.dist.js'
-export { EnvelopeIcon } from './EnvelopeIcon.dist.js'
-export { ExternalLinkIcon } from './ExternalLinkIcon.dist.js'
-export { EyeIcon } from './EyeIcon.dist.js'
-export { EyeOffIcon } from './EyeOffIcon.dist.js'
-export { FileIcon } from './FileIcon.dist.js'
-export { FilterIcon } from './FilterIcon.dist.js'
-export { FullscreenExitIcon } from './FullscreenExitIcon.dist.js'
-export { FullscreenIcon } from './FullscreenIcon.dist.js'
-export { GearIcon } from './GearIcon.dist.js'
-export { GlobeIcon } from './GlobeIcon.dist.js'
-export { HandIcon } from './HandIcon.dist.js'
-export { HelpIcon } from './HelpIcon.dist.js'
-export { HistoryIcon } from './HistoryIcon.dist.js'
-export { HomeAdminToolsIcon } from './HomeAdminToolsIcon.dist.js'
-export { HomeContentToolsIcon } from './HomeContentToolsIcon.dist.js'
-export { HomeIcon } from './HomeIcon.dist.js'
-export { InfoIcon } from './InfoIcon.dist.js'
-export { InteractiveIcon } from './InteractiveIcon.dist.js'
-export { IqIcon } from './IqIcon.dist.js'
-export { LabsIcon } from './LabsIcon.dist.js'
-export { LinkIcon } from './LinkIcon.dist.js'
-export { ListIcon } from './ListIcon.dist.js'
-export { LocationIcon } from './LocationIcon.dist.js'
-export { LockFillIcon } from './LockFillIcon.dist.js'
-export { LockIcon } from './LockIcon.dist.js'
-export { MegaphoneIcon } from './MegaphoneIcon.dist.js'
-export { MenuCloseIcon } from './MenuCloseIcon.dist.js'
-export { MenuIcon } from './MenuIcon.dist.js'
-export { MicIcon } from './MicIcon.dist.js'
-export { MoreIcon } from './MoreIcon.dist.js'
-export { MoveIcon } from './MoveIcon.dist.js'
-export { NextIcon } from './NextIcon.dist.js'
-export { NoteIcon } from './NoteIcon.dist.js'
-export { NotificationsIcon } from './NotificationsIcon.dist.js'
-export { NotificationsNewIcon } from './NotificationsNewIcon.dist.js'
-export { ObjectiveIcon } from './ObjectiveIcon.dist.js'
-export { OrgIcon } from './OrgIcon.dist.js'
-export { PaperclipIcon } from './PaperclipIcon.dist.js'
-export { PathIcon } from './PathIcon.dist.js'
-export { PauseIcon } from './PauseIcon.dist.js'
-export { PencilIcon } from './PencilIcon.dist.js'
-export { PeopleIcon } from './PeopleIcon.dist.js'
-export { PhoneDownIcon } from './PhoneDownIcon.dist.js'
-export { PhoneIcon } from './PhoneIcon.dist.js'
-export { PlaceholderIcon } from './PlaceholderIcon.dist.js'
-export { PlayCircleIcon } from './PlayCircleIcon.dist.js'
-export { PlayIcon } from './PlayIcon.dist.js'
-export { PlusCircleFillIcon } from './PlusCircleFillIcon.dist.js'
-export { PlusCircleIcon } from './PlusCircleIcon.dist.js'
-export { PlusIcon } from './PlusIcon.dist.js'
-export { PodcastIcon } from './PodcastIcon.dist.js'
-export { PreviousIcon } from './PreviousIcon.dist.js'
-export { ProjectIcon } from './ProjectIcon.dist.js'
-export { QNAIcon } from './QNAIcon.dist.js'
-export { QuestionFilledIcon } from './QuestionFilledIcon.dist.js'
-export { QuestionIcon } from './QuestionIcon.dist.js'
-export { RedoIcon } from './RedoIcon.dist.js'
-export { ReloadIcon } from './ReloadIcon.dist.js'
-export { ReportIcon } from './ReportIcon.dist.js'
-export { ScopeIcon } from './ScopeIcon.dist.js'
-export { ScreenshareIcon } from './ScreenshareIcon.dist.js'
-export { SearchIcon } from './SearchIcon.dist.js'
-export { ServiceBoxIcon } from './ServiceBoxIcon.dist.js'
-export { ServiceDropboxIcon } from './ServiceDropboxIcon.dist.js'
-export { ServiceGoogleDriveIcon } from './ServiceGoogleDriveIcon.dist.js'
-export { ShareIcon } from './ShareIcon.dist.js'
-export { SignoutIcon } from './SignoutIcon.dist.js'
-export { SkipBackwardIcon } from './SkipBackwardIcon.dist.js'
-export { SkipForwardIcon } from './SkipForwardIcon.dist.js'
-export { SocialDribbbleIcon } from './SocialDribbbleIcon.dist.js'
-export { SocialFacebookIcon } from './SocialFacebookIcon.dist.js'
-export { SocialGithubIcon } from './SocialGithubIcon.dist.js'
-export { SocialGooglePlusIcon } from './SocialGooglePlusIcon.dist.js'
-export { SocialLinkedinIcon } from './SocialLinkedinIcon.dist.js'
-export { SocialMediumIcon } from './SocialMediumIcon.dist.js'
-export { SocialRssIcon } from './SocialRssIcon.dist.js'
-export { SocialStackoverflowIcon } from './SocialStackoverflowIcon.dist.js'
-export { SocialTwitterIcon } from './SocialTwitterIcon.dist.js'
-export { SortAscIcon } from './SortAscIcon.dist.js'
-export { SortDescIcon } from './SortDescIcon.dist.js'
-export { SortIcon } from './SortIcon.dist.js'
-export { StarFillIcon } from './StarFillIcon.dist.js'
-export { StarIcon } from './StarIcon.dist.js'
-export { StopIcon } from './StopIcon.dist.js'
-export { TagAssignmentIcon } from './TagAssignmentIcon.dist.js'
-export { TagIcon } from './TagIcon.dist.js'
-export { TagManagementIcon } from './TagManagementIcon.dist.js'
-export { TestIcon } from './TestIcon.dist.js'
-export { ThemeIcon } from './ThemeIcon.dist.js'
-export { ThumbDownFillIcon } from './ThumbDownFillIcon.dist.js'
-export { ThumbDownIcon } from './ThumbDownIcon.dist.js'
-export { ThumbUpFillIcon } from './ThumbUpFillIcon.dist.js'
-export { ThumbUpIcon } from './ThumbUpIcon.dist.js'
-export { TranscriptIcon } from './TranscriptIcon.dist.js'
-export { TrashIcon } from './TrashIcon.dist.js'
-export { UndoIcon } from './UndoIcon.dist.js'
-export { UnionIcon } from './UnionIcon.dist.js'
-export { UploadIcon } from './UploadIcon.dist.js'
-export { UserAddIcon } from './UserAddIcon.dist.js'
-export { UserIcon } from './UserIcon.dist.js'
-export { VideoIcon } from './VideoIcon.dist.js'
-export { VolumeHighIcon } from './VolumeHighIcon.dist.js'
-export { VolumeLowIcon } from './VolumeLowIcon.dist.js'
-export { VolumeMediumIcon } from './VolumeMediumIcon.dist.js'
-export { VolumeOffIcon } from './VolumeOffIcon.dist.js'
-export { WarningFilledIcon } from './WarningFilledIcon.dist.js'
-export { WarningIcon } from './WarningIcon.dist.js'
-export { WindowIcon } from './WindowIcon.dist.js'
-export { XCircleFillIcon } from './XCircleFillIcon.dist.js'
-export { XCircleIcon } from './XCircleIcon.dist.js'
