+++ packages/icon/src/react/icons/index.tsx
@@ -1,0 +1,163 @@
+export { AccountIcon } from './AccountIcon.dist'
+export { AnalyticsIcon } from './AnalyticsIcon.dist'
+export { ArrowDownIcon } from './ArrowDownIcon.dist'
+export { ArrowDownRightIcon } from './ArrowDownRightIcon.dist'
+export { ArrowLeftIcon } from './ArrowLeftIcon.dist'
+export { ArrowRightIcon } from './ArrowRightIcon.dist'
+export { ArrowUpIcon } from './ArrowUpIcon.dist'
+export { ArrowUpRightIcon } from './ArrowUpRightIcon.dist'
+export { ArticleIcon } from './ArticleIcon.dist'
+export { AuthorKitIcon } from './AuthorKitIcon.dist'
+export { AuthorsNestIcon } from './AuthorsNestIcon.dist'
+export { BellIcon } from './BellIcon.dist'
+export { BellRungIcon } from './BellRungIcon.dist'
+export { BookmarkFillIcon } from './BookmarkFillIcon.dist'
+export { BookmarkIcon } from './BookmarkIcon.dist'
+export { BriefcaseIcon } from './BriefcaseIcon.dist'
+export { BrowseIcon } from './BrowseIcon.dist'
+export { BuildingIcon } from './BuildingIcon.dist'
+export { CalendarIcon } from './CalendarIcon.dist'
+export { CaretAltDownIcon } from './CaretAltDownIcon.dist'
+export { CaretAltLeftIcon } from './CaretAltLeftIcon.dist'
+export { CaretAltRightIcon } from './CaretAltRightIcon.dist'
+export { CaretAltUpIcon } from './CaretAltUpIcon.dist'
+export { CaretDownIcon } from './CaretDownIcon.dist'
+export { CaretLeftIcon } from './CaretLeftIcon.dist'
+export { CaretRightIcon } from './CaretRightIcon.dist'
+export { CaretUpIcon } from './CaretUpIcon.dist'
+export { ChannelAddIcon } from './ChannelAddIcon.dist'
+export { ChannelIcon } from './ChannelIcon.dist'
+export { ChatAltFillIcon } from './ChatAltFillIcon.dist'
+export { ChatAltIcon } from './ChatAltIcon.dist'
+export { ChatFillIcon } from './ChatFillIcon.dist'
+export { ChatIcon } from './ChatIcon.dist'
+export { CheckCircleFillIcon } from './CheckCircleFillIcon.dist'
+export { CheckCircleIcon } from './CheckCircleIcon.dist'
+export { CheckIcon } from './CheckIcon.dist'
+export { ClockIcon } from './ClockIcon.dist'
+export { ClockWarningIcon } from './ClockWarningIcon.dist'
+export { CloseCaptionedFillIcon } from './CloseCaptionedFillIcon.dist'
+export { CloseCaptionedIcon } from './CloseCaptionedIcon.dist'
+export { CloseIcon } from './CloseIcon.dist'
+export { CloudIcon } from './CloudIcon.dist'
+export { CloudOffIcon } from './CloudOffIcon.dist'
+export { CodeBrandedIcon } from './CodeBrandedIcon.dist'
+export { CodeIcon } from './CodeIcon.dist'
+export { ConnectionWarningIcon } from './ConnectionWarningIcon.dist'
+export { CopyIcon } from './CopyIcon.dist'
+export { CoursesIcon } from './CoursesIcon.dist'
+export { CurriculumIcon } from './CurriculumIcon.dist'
+export { DashboardIcon } from './DashboardIcon.dist'
+export { DesktopIcon } from './DesktopIcon.dist'
+export { DownloadIcon } from './DownloadIcon.dist'
+export { EnvelopeArrowIcon } from './EnvelopeArrowIcon.dist'
+export { EnvelopeIcon } from './EnvelopeIcon.dist'
+export { ExternalLinkIcon } from './ExternalLinkIcon.dist'
+export { EyeIcon } from './EyeIcon.dist'
+export { EyeOffIcon } from './EyeOffIcon.dist'
+export { FileIcon } from './FileIcon.dist'
+export { FilterIcon } from './FilterIcon.dist'
+export { FullscreenExitIcon } from './FullscreenExitIcon.dist'
+export { FullscreenIcon } from './FullscreenIcon.dist'
+export { GearIcon } from './GearIcon.dist'
+export { GlobeIcon } from './GlobeIcon.dist'
+export { HandIcon } from './HandIcon.dist'
+export { HelpIcon } from './HelpIcon.dist'
+export { HistoryIcon } from './HistoryIcon.dist'
+export { HomeAdminToolsIcon } from './HomeAdminToolsIcon.dist'
+export { HomeContentToolsIcon } from './HomeContentToolsIcon.dist'
+export { HomeIcon } from './HomeIcon.dist'
+export { InfoIcon } from './InfoIcon.dist'
+export { InteractiveIcon } from './InteractiveIcon.dist'
+export { IqIcon } from './IqIcon.dist'
+export { LabsIcon } from './LabsIcon.dist'
+export { LinkIcon } from './LinkIcon.dist'
+export { ListIcon } from './ListIcon.dist'
+export { LocationIcon } from './LocationIcon.dist'
+export { LockFillIcon } from './LockFillIcon.dist'
+export { LockIcon } from './LockIcon.dist'
+export { MegaphoneIcon } from './MegaphoneIcon.dist'
+export { MenuCloseIcon } from './MenuCloseIcon.dist'
+export { MenuIcon } from './MenuIcon.dist'
+export { MicIcon } from './MicIcon.dist'
+export { MoreIcon } from './MoreIcon.dist'
+export { MoveIcon } from './MoveIcon.dist'
+export { NextIcon } from './NextIcon.dist'
+export { NoteIcon } from './NoteIcon.dist'
+export { NotificationsIcon } from './NotificationsIcon.dist'
+export { NotificationsNewIcon } from './NotificationsNewIcon.dist'
+export { ObjectiveIcon } from './ObjectiveIcon.dist'
+export { OrgIcon } from './OrgIcon.dist'
+export { PaperclipIcon } from './PaperclipIcon.dist'
+export { PathIcon } from './PathIcon.dist'
+export { PauseIcon } from './PauseIcon.dist'
+export { PencilIcon } from './PencilIcon.dist'
+export { PeopleIcon } from './PeopleIcon.dist'
+export { PhoneDownIcon } from './PhoneDownIcon.dist'
+export { PhoneIcon } from './PhoneIcon.dist'
+export { PlaceholderIcon } from './PlaceholderIcon.dist'
+export { PlayCircleIcon } from './PlayCircleIcon.dist'
+export { PlayIcon } from './PlayIcon.dist'
+export { PlusCircleFillIcon } from './PlusCircleFillIcon.dist'
+export { PlusCircleIcon } from './PlusCircleIcon.dist'
+export { PlusIcon } from './PlusIcon.dist'
+export { PodcastIcon } from './PodcastIcon.dist'
+export { PreviousIcon } from './PreviousIcon.dist'
+export { ProjectIcon } from './ProjectIcon.dist'
+export { QNAIcon } from './QNAIcon.dist'
+export { QuestionFilledIcon } from './QuestionFilledIcon.dist'
+export { QuestionIcon } from './QuestionIcon.dist'
+export { RedoIcon } from './RedoIcon.dist'
+export { ReloadIcon } from './ReloadIcon.dist'
+export { ReportIcon } from './ReportIcon.dist'
+export { ScopeIcon } from './ScopeIcon.dist'
+export { ScreenshareIcon } from './ScreenshareIcon.dist'
+export { SearchIcon } from './SearchIcon.dist'
+export { ServiceBoxIcon } from './ServiceBoxIcon.dist'
+export { ServiceDropboxIcon } from './ServiceDropboxIcon.dist'
+export { ServiceGoogleDriveIcon } from './ServiceGoogleDriveIcon.dist'
+export { ShareIcon } from './ShareIcon.dist'
+export { SignoutIcon } from './SignoutIcon.dist'
+export { SkipBackwardIcon } from './SkipBackwardIcon.dist'
+export { SkipForwardIcon } from './SkipForwardIcon.dist'
+export { SocialDribbbleIcon } from './SocialDribbbleIcon.dist'
+export { SocialFacebookIcon } from './SocialFacebookIcon.dist'
+export { SocialGithubIcon } from './SocialGithubIcon.dist'
+export { SocialGooglePlusIcon } from './SocialGooglePlusIcon.dist'
+export { SocialLinkedinIcon } from './SocialLinkedinIcon.dist'
+export { SocialMediumIcon } from './SocialMediumIcon.dist'
+export { SocialRssIcon } from './SocialRssIcon.dist'
+export { SocialStackoverflowIcon } from './SocialStackoverflowIcon.dist'
+export { SocialTwitterIcon } from './SocialTwitterIcon.dist'
+export { SortAscIcon } from './SortAscIcon.dist'
+export { SortDescIcon } from './SortDescIcon.dist'
+export { SortIcon } from './SortIcon.dist'
+export { StarFillIcon } from './StarFillIcon.dist'
+export { StarIcon } from './StarIcon.dist'
+export { StopIcon } from './StopIcon.dist'
+export { TagAssignmentIcon } from './TagAssignmentIcon.dist'
+export { TagIcon } from './TagIcon.dist'
+export { TagManagementIcon } from './TagManagementIcon.dist'
+export { TestIcon } from './TestIcon.dist'
+export { ThemeIcon } from './ThemeIcon.dist'
+export { ThumbDownFillIcon } from './ThumbDownFillIcon.dist'
+export { ThumbDownIcon } from './ThumbDownIcon.dist'
+export { ThumbUpFillIcon } from './ThumbUpFillIcon.dist'
+export { ThumbUpIcon } from './ThumbUpIcon.dist'
+export { TranscriptIcon } from './TranscriptIcon.dist'
+export { TrashIcon } from './TrashIcon.dist'
+export { UndoIcon } from './UndoIcon.dist'
+export { UnionIcon } from './UnionIcon.dist'
+export { UploadIcon } from './UploadIcon.dist'
+export { UserAddIcon } from './UserAddIcon.dist'
+export { UserIcon } from './UserIcon.dist'
+export { VideoIcon } from './VideoIcon.dist'
+export { VolumeHighIcon } from './VolumeHighIcon.dist'
+export { VolumeLowIcon } from './VolumeLowIcon.dist'
+export { VolumeMediumIcon } from './VolumeMediumIcon.dist'
+export { VolumeOffIcon } from './VolumeOffIcon.dist'
+export { WarningFilledIcon } from './WarningFilledIcon.dist'
+export { WarningIcon } from './WarningIcon.dist'
+export { WindowIcon } from './WindowIcon.dist'
+export { XCircleFillIcon } from './XCircleFillIcon.dist'
+export { XCircleIcon } from './XCircleIcon.dist'
