+++ packages/icon/src/vars/index.ts
